+++ apps/backend/src/libs/plugins/authorization-plugin/authorization-plugin.ts
@@ -8,7 +8,6 @@ import { AuthError } from "~/modules/auth/auth.js";
 import { type UserService } from "~/modules/users/users.js";
 
 import { ServerHooks } from "../libs/enums/enums.js";
-import { ONE_THOUSAND_MILLISECONDS } from "./libs/constants/constants.js";
 import { checkIsWhiteRoute } from "./libs/helpers/helpers.js";
 
 type PluginOptions = {
@@ -37,14 +36,8 @@ const authorizationPlugin = fp<PluginOptions>((app, options, done) => {
 
 		try {
 			const {
-				payload: { exp, userId },
+				payload: { userId },
 			} = await token.decode(headerToken);
-
-			if (
-				(exp as number) < Math.floor(Date.now() / ONE_THOUSAND_MILLISECONDS)
-			) {
-				throw new AuthError({ message: ErrorMessage.UNAUTHORIZED });
-			}
 
 			const user = await userService.find(userId);