+++ apps/frontend/src/libs/modules/store/libs/middlewares/middlewares.ts
@@ -1,0 +1,1 @@
+export { handleErrorMiddleware } from "./handle-error.middleware.js";