+++ apps/frontend/src/libs/modules/store/store.module.ts
@@ -9,6 +9,7 @@ import { AppEnvironment } from "~/libs/enums/enums.js";
 import { type Config } from "~/libs/modules/config/config.js";
 import { notification } from "~/libs/modules/notification/notification.js";
 import { storage } from "~/libs/modules/storage/storage.js";
+import { reducer as appReducer } from "~/modules/app/app.js";
 import { authApi, reducer as authReducer } from "~/modules/auth/auth.js";
 import {
 	onboardingApi,
@@ -18,9 +19,10 @@ import { quizApi, reducer as quizReducer } from "~/modules/quiz/quiz.js";
 import { tasksApi, reducer as tasksReducer } from "~/modules/tasks/tasks.js";
 import { usersApi, reducer as usersReducer } from "~/modules/users/users.js";
 
-import { handleErrorMiddleware } from "./handle-error.middleware.js";
+import { handleErrorMiddleware } from "./libs/middlewares/middlewares.js";
 
 type RootReducer = {
+	app: ReturnType<typeof appReducer>;
 	auth: ReturnType<typeof authReducer>;
 	onboarding: ReturnType<typeof onboardingReducer>;
 	quiz: ReturnType<typeof quizReducer>;
@@ -55,9 +57,10 @@ class Store {
 					thunk: {
 						extraArgument: this.extraArguments,
 					},
-				}).prepend([handleErrorMiddleware]);
+				}).prepend([handleErrorMiddleware(this.extraArguments)]);
 			},
 			reducer: {
+				app: appReducer,
 				auth: authReducer,
 				onboarding: onboardingReducer,
 				quiz: quizReducer,
@@ -80,4 +83,4 @@ class Store {
 	}
 }
 
-export { Store };
+export { type ExtraArguments, Store };