+++ apps/mobile/src/packages/users/libs/types/types.ts
@@ -1,6 +1,6 @@
 export {
 	type UserDto,
-	type UserGetAllResponseDto,
+	type UserGetParametersDto,
 	type UserSignInRequestDto,
 	type UserSignInResponseDto,
 	type UserSignUpFormDto,